Refactor based on code review - WIP
