+++ src/lib/lowerCaseArray.js
@@ -1,0 +1,3 @@
+const lowerCaseArray = (array) => array.map((item) => item.toLowerCase())
+
+module.exports = lowerCaseArray;
