+++ src/test/unit.test.js
@@ -3,8 +3,9 @@ const test = require('tape');
 const { addUserStatus } = require('../controllers/middleware');
 const { jobStatusText, jobPrefIsPublic } = require('../views/helpers/index');
 const makeFacCohortName = require('../lib/makeFacCohortName');
-const remDupsFromArray = require('../lib/remDupsFromArray');
+const removeArrayDuplicates = require('../lib/removeArrayDuplicates');
 const githubApiCalls = require('../lib/githubApiCalls');
+const lowerCaseArray = require('../lib/lowerCaseArray');
 
 
 test('UNIT TESTS', (t) => {
@@ -93,9 +94,17 @@ test('Test makeFacCohortName', (t) => {
 });
 
 // DUPLICATE HELPER in Lib folder
-test('Test remDupsFromArray', (t) => {
+test('Test removeArrayDuplicates', (t) => {
   const arrayInput = ['javascript', 'javascript', 'JavaScript', 'Node.js']
   const expected = ['javascript', 'Node.js']
-  t.deepEquals(remDupsFromArray(arrayInput), expected, 'removed');
+  t.deepEquals(removeArrayDuplicates(arrayInput), expected, 'removed');
+  t.end();
+});
+
+// LOWERCASE HELPER in Lib folder
+test('Test lowerCaseArray', (t) => {
+  const arrayInput = ['JAVASCRIPT', 'javascript', 'JavaScript', 'Node.js'];
+  const expected = ['javascript', 'javascript', 'javascript', 'node.js'];
+  t.deepEquals(lowerCaseArray(arrayInput), expected, 'changed to lowercase');
   t.end();
 });
